riscv_ctg/requirements.txt
click
ruamel.yaml>=0.16.0
colorlog
pytest
